requirements.txt
# Существующие зависимости
aiogram>=3.0.0
networkx>=3.0
numpy>=1.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
langdetect>=1.0.9
deep-translator>=1.11.4
aiohttp>=3.8.0
asyncio>=3.4.3
requests>=2.28.0

# Новые зависимости для мультимодальности
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pandas>=2.0.0
PyYAML>=6.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Опционально (ML модели - тяжелые)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ultralytics>=8.0.0
easyocr>=1.7.0
openai-whisper>=20231117